refactor: moved styles
