+++ client/src/shared/ui/buttons/logout-button/index.ts
@@ -1,1 +1,1 @@
-export { LogoutButton } from './logout-button';
+export { LogoutButton } from './ui/logout-button';
