+++ client/src/widgets/layout/ui/footer/ui/footer.styles.ts
@@ -1,0 +1,8 @@
+import { Layout } from 'antd';
+import styled from 'styled-components';
+
+export const StyledFooter = styled(Layout.Footer)`
+  background-color: var(--main-background-color);
+  z-index: var(--footer-z-index);
+  height: var(--footer-height);
+`;
