+++ client/src/widgets/layout/ui/footer/ui/footer.tsx
@@ -1,12 +1,5 @@
-import { Layout } from 'antd';
-import styled from 'styled-components';
+import { StyledFooter } from './footer.styles';
 
 export const Footer = () => {
   return <StyledFooter>Footer</StyledFooter>;
 };
-
-const StyledFooter = styled(Layout.Footer)`
-  background-color: var(--main-background-color);
-  z-index: var(--footer-z-index);
-  height: var(--footer-height);
-`;
